requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
qrcode[pil]==8.2
Pillow>=9.1.0
Flask==2.3.2
user-agents==2.2.0
requests==2.31.0
